FEAT: added code for the sensor
mostly testing stuff. Later we can alter it so it does something useful, such as sending it
to the backend.

Something wrong with the connections in the sensor. Maybe the cables.
Maybe its this fucked up esp i got. Anyway, we got the code but
we're unable to read the blasted sensor. Not great.